Defend against Prototype.js
When testing code that needs to run on a page that also has Prototype.js 1.5.1 I discovered that Prototype breaks `sendMessage` in `bridge.js` because Prototype < 1.7.1 extended native objects in an unsafe way.

Since `sendMessage` will alway attempt to stringify an `Array` the `stringify` function added to `bridge.js` should be relatively safe.